pyproject.toml
[project]
name = "telellm"
version = "0.1.0"
description = "Telegram Bot with LLM integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.17.0",
    "loguru>=0.7.3",
    "pytest>=8.3.3",
    "python-dotenv>=1.0.1",
    "redis>=5.2.1",
]

[dependency-groups]
dev = [
    "ruff>=0.7.3",
]
